refactor floating point calculation

+++ pd/include/ui/solver_params.h
@@ -13,7 +13,7 @@ namespace ui
 
 	struct SolverParams
 	{
-		float dt{ 0.0166667f };
+		double dt{ 0.0166667 }; // use double precision to avoid floating point numeric error
 		int n_solver_pd_iterations{ 10 };
 		int n_itr_solver_iterations{ 500 };
